Add util.ts
